Fix(summary)!: Reformat Summary CSV file
This commit changes the results summary file to be a well-formatted csv
file.  Previously the summary csv was missing test_type, packet_type and
packet_size headers, and there are multiple tables embedded to handle
throughput, iops and latency.  This commit adds the missing headers, and
consolidates throughput, iops and latency into the same table.  This is
to help with CSV to JSON conversion.

Co-developed-by: Cursor